backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
python-docx==1.1.2
reportlab==4.2.2
requests==2.32.5
python-dotenv==1.0.1
pypdf==5.4.0
pymupdf==1.24.10
pydantic==2.9.2
fonttools==4.54.1
Pillow==10.4.0
numpy==2.1.2
psycopg[binary]==3.2.3
pytest==8.3.3
httpx==0.27.2
